Fixed long-query in generating Archive feed
This addresses a problem with a query used to generate the archival feed
read by Merritt (repository). The query performed very badly. It wasn't
easy to notice until the Merritt made requests pretty far into the list
of resources to be accessioned. It turns out the query wasn't using the
correct index and so it was taking about 2 seconds to execute. Given the
other traffic Open Context responds to, these 2 second delays started
adding up, causing some tables to lock and others to crash entirely.

The query is now forced use the appropriate index, the WHERE condition
also seems to help speed things up.